Rename `LiveDataMapper` to `HackrNewsFeedMapper`
