+++ HackrNews/Live Hackr News API/RemoteLiveHackrNewsLoader.swift
@@ -8,6 +8,6 @@ public typealias RemoteLiveHackrNewsLoader = RemoteLoader<[LiveHackrNew]>
 
 public extension RemoteLiveHackrNewsLoader {
     convenience init(url: URL, client: HTTPClient) {
-        self.init(url: url, client: client, mapper: LiveDataMapper.map)
+        self.init(url: url, client: client, mapper: HackrNewsFeedMapper.map)
     }
 }
